# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.DisciplinaMatriculaCreateDTO:
    properties:
      aluno_unique_id:
        type: integer
      curso_disciplina_unique_i:
        type: integer
    required:
    - aluno_unique_id
    - curso_disciplina_unique_i
    type: object
  dtos.DisciplinaMatriculaUpdateDTO:
    properties:
      aluno_unique_id:
        type: integer
      curso_disciplina_unique_i:
        type: integer
    type: object
...
  dtos.FaculdadeCreateDTO:
    properties:
      cnpj:
        type: string
      nome:
        type: string
    required:
    - cnpj
    - nome
    type: object
  dtos.FaculdadeUpdateDTO:
    properties:
      cnpj:
        type: string
      nome:
        type: string
    type: object
...
  models.Aluno:
    properties:
      cpf:
        type: string
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cpf
    - nome
    type: object
  models.Curso:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      faculdade:
        $ref: '#/definitions/models.Faculdade'
      faculdade_unique_id:
        type: integer
      nome:
        type: string
      unique_id:
        type: integer
    type: object
  models.CursoDisciplina:
    properties:
      curso:
        $ref: '#/definitions/models.Curso'
      curso_unique_id:
        type: integer
      disciplina:
        $ref: '#/definitions/models.Disciplina'
      disciplina_matriculas:
        items:
          $ref: '#/definitions/models.DisciplinaMatricula'
        type: array
      disciplina_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.Disciplina:
    properties:
      curso_disciplinas:
        items:
          $ref: '#/definitions/models.CursoDisciplina'
        type: array
      nome:
        type: string
      professor:
        $ref: '#/definitions/models.Professor'
      professor_unique_id:
        type: integer
      unique_id:
        type: integer
    type: object
  models.DisciplinaMatricula:
    properties:
      aluno:
        $ref: '#/definitions/models.Aluno'
      aluno_unique_id:
        type: integer
      curso_disciplina:
        $ref: '#/definitions/models.CursoDisciplina'
      curso_disciplina_unique_id:
        type: integer
      data_matricula:
        type: string
      unique_id:
        type: integer
    type: object
  models.Faculdade:
    properties:
      cnpj:
        type: string
      cursos:
        items:
          $ref: '#/definitions/models.Curso'
        type: array
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - cnpj
    - nome
    type: object
  models.Professor:
    properties:
      disciplinas:
        items:
          $ref: '#/definitions/models.Disciplina'
        type: array
      formacao:
        type: string
      nome:
        type: string
      unique_id:
        type: integer
    required:
    - formacao
    - nome
    type: object
host: localhost:5000
info:
  contact:
    name: Tathiane Souza
  description: Esta é a documentação API em Golang
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API Faculdade passei direto
  version: "1.0"
paths:
...
  /disciplina-matricula:
    get:
      consumes:
      - application/json
      description: Busca todas as disciplinas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DisciplinaMatricula'
      summary: Buscar todas as disciplinas
      tags:
      - Controller DisciplinaMatricula
    post:
      consumes:
      - application/json
      description: 'Cria uma disciplinaMatricula, os campos que devem ser enviados
        são: aluno_unique_id e curso_disciplina_unique_id'
      parameters:
      - description: disciplinaMatricula
        in: body
        name: disciplinaMatricula
        required: true
        schema:
          $ref: '#/definitions/dtos.DisciplinaMatriculaCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DisciplinaMatricula'
      summary: Cria uma disciplinaMatricula
      tags:
      - Controller DisciplinaMatricula
  /disciplina-matricula/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta uma disciplinaMatricula
      parameters:
      - description: UniqueID da disciplinaMatricula
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta uma disciplinaMatricula
      tags:
      - Controller DisciplinaMatricula
    get:
      consumes:
      - application/json
      description: Busca disciplinaMatricula por id
      parameters:
      - description: UniqueID do disciplinaMatricula
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DisciplinaMatricula'
      summary: Busca disciplinaMatricula por id
      tags:
      - Controller DisciplinaMatricula
    put:
      consumes:
      - application/json
      description: 'Edita uma disciplinaMatricula, os campos que podem ser enviados
        são: aluno_unique_id e curso_disciplina_unique_id'
      parameters:
      - description: UniqueID da disciplinaMatricula
        in: path
        name: id
        required: true
        type: integer
      - description: disciplinaMatricula
        in: body
        name: disciplinaMatricula
        required: true
        schema:
          $ref: '#/definitions/dtos.DisciplinaMatriculaUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DisciplinaMatricula'
      summary: Edita uma disciplinaMatricula
      tags:
      - Controller DisciplinaMatricula
...
  /faculdade:
    get:
      consumes:
      - application/json
      description: Busca todas as faculdades
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Faculdade'
      summary: Buscar todas as faculdades
      tags:
      - Controller Faculdade
    post:
      consumes:
      - application/json
      description: 'Cria uma faculdade, os campos que devem ser enviados são: nome
        e cnpj'
      parameters:
      - description: faculdade
        in: body
        name: faculdade
        required: true
        schema:
          $ref: '#/definitions/dtos.FaculdadeCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Faculdade'
      summary: Cria uma faculdade
      tags:
      - Controller Faculdade
  /faculdade/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta uma faculdade
      parameters:
      - description: UniqueID da faculdade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta uma faculdade
      tags:
      - Controller Faculdade
    get:
      consumes:
      - application/json
      description: Busca faculdade por id
      parameters:
      - description: UniqueID do faculdade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Faculdade'
      summary: Busca faculdade por id
      tags:
      - Controller Faculdade
    put:
      consumes:
      - application/json
      description: 'Edita uma faculdade, os campos que devem ser enviados são: nome
        e cnpj'
      parameters:
      - description: UniqueID da faculdade
        in: path
        name: id
        required: true
        type: integer
      - description: faculdade
        in: body
        name: faculdade
        required: true
        schema:
          $ref: '#/definitions/dtos.FaculdadeUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Faculdade'
      summary: Edita uma faculdade
      tags:
      - Controller Faculdade
...
swagger: "2.0"